+++ Bugsnag/Assets/Bugsnag/Runtime/Delivery.cs
@@ -67,6 +67,8 @@ namespace BugsnagUnity
                 var report = (Report)payload;
                 if (_configuration.GetOnSendErrorCallbacks().Count > 0)
                 {
+                    // save handled state before callbacks so we can check later if it was overidden
+                    var initialUnhandledState = report.Event.Unhandled;
                     lock (_callbackLock)
                     {
                         foreach (var onSendErrorCallback in _configuration.GetOnSendErrorCallbacks())
@@ -84,6 +86,10 @@ namespace BugsnagUnity
                             }
                         }
                     }
+                    if (initialUnhandledState != report.Event.Unhandled)
+                    {
+                        report.Event.UnhandledOverridden();
+                    }
                 }
                 report.Event.RedactMetadata(_configuration);
                 //pipeline expects and array of events even though we only ever report 1
@@ -91,7 +97,7 @@ namespace BugsnagUnity
             }
             try
             {
-                MainThreadDispatchBehaviour.Instance().Enqueue(PushToServer(payload));
+                MainThreadDispatchBehaviour.Enqueue(PushToServer(payload));
             }
             catch
             {
@@ -110,7 +116,8 @@ namespace BugsnagUnity
             else
             {
                 var networkCheckDone = false;
-                new Thread(() => {
+                new Thread(() =>
+                {
                     shouldDeliver = _client.NativeClient.ShouldAttemptDelivery();
                     networkCheckDone = true;
                 }).Start();
@@ -158,7 +165,8 @@ namespace BugsnagUnity
                 else
                 {
                     var bodyReady = false;
-                    new Thread(() => {
+                    new Thread(() =>
+                    {
                         try
                         {
                             body = PrepareEventBody(payload);
@@ -210,7 +218,7 @@ namespace BugsnagUnity
                     _payloadManager.RemovePayload(payload);
                     StartDeliveringCachedPayloads();
                 }
-                else if ( code == 0 || code == 408 || code == 429 || code >= 500)
+                else if (code == 0 || code == 408 || code == 429 || code >= 500)
                 {
                     // sending failed with no network or retryable error, cache payload to disk
                     _payloadManager.SendPayloadFailed(payload);
@@ -250,7 +258,7 @@ namespace BugsnagUnity
                     }
                 }
             }
-           
+
             return serialisedPayload;
         }
 
@@ -277,7 +285,7 @@ namespace BugsnagUnity
                     }
                 }
             }
-           
+
             return serialisedPayload;
         }
 
@@ -508,7 +516,7 @@ namespace BugsnagUnity
             try
             {
                 _finishedCacheDeliveries.Clear();
-                MainThreadDispatchBehaviour.Instance().Enqueue(DeliverCachedPayloads());
+                MainThreadDispatchBehaviour.Enqueue(DeliverCachedPayloads());
             }
             catch
             {
@@ -587,7 +595,7 @@ namespace BugsnagUnity
                         }
                         Deliver(report);
                     }
-                    
+
                     yield return new WaitUntil(() => CachedPayloadProcessed(id));
                 }
             }